+++ src/components/ItemForm/ItemForm.js
@@ -32,18 +32,19 @@ class ItemForm extends React.Component {
     onOverride: PropTypes.func.isRequired,
     form: PropTypes.object.isRequired,
     formRef: PropTypes.object.isRequired,
-    checkoutError: PropTypes.object,
+    checkoutError: PropTypes.arrayOf(
+      PropTypes.object
+    ).isRequired,
     onClearCheckoutErrors: PropTypes.func,
   };
 
   static defaultProps = {
     patron: {},
     item: {},
-    checkoutError: {},
   };
 
-  static getDerivedStateFromProps(props, { error }) {
-    return { error: props.checkoutError?.item || error || {} };
+  static getDerivedStateFromProps(props, { errors }) {
+    return { errors: props.checkoutError || errors || [] };
   }
 
   constructor(props) {
@@ -52,7 +53,7 @@ class ItemForm extends React.Component {
     this.barcodeEl = React.createRef();
     this.state = {
       overrideModalOpen: false,
-      error: {},
+      errors: [],
     };
   }
 
@@ -84,8 +85,8 @@ class ItemForm extends React.Component {
     }
   }
 
-  setError = (error) => {
-    this.setState({ error });
+  setError = (errors) => {
+    this.setState({ errors });
   };
 
   openOverrideModal = () => {
@@ -106,17 +107,17 @@ class ItemForm extends React.Component {
       onClearCheckoutErrors,
     } = this.props;
 
-    this.setError({});
+    this.setError([]);
     onClearCheckoutErrors();
     form.reset();
   };
 
   onSubmit = async (event) => {
     const { handleSubmit } = this.props;
-    const error = await handleSubmit(event);
+    const errors = await handleSubmit(event);
 
-    if (error?.item) {
-      return error;
+    if (!isEmpty(errors)) {
+      return errors;
     }
 
     return this.clearForm();
@@ -130,7 +131,7 @@ class ItemForm extends React.Component {
       onOverride,
     } = this.props;
 
-    const { error, overrideModalOpen } = this.state;
+    const { errors, overrideModalOpen } = this.state;
     const validationEnabled = false;
 
     return (
@@ -173,13 +174,12 @@ class ItemForm extends React.Component {
           </Row>
         </form>
         {
-          !isEmpty(error) &&
+          !isEmpty(errors) &&
           <ErrorModal
             stripes={stripes}
             item={item || {}}
-            message={error.barcode}
-            loanPolicy={error.loanPolicy}
-            open={!isEmpty(error)}
+            errors={errors}
+            open
             openOverrideModal={this.openOverrideModal}
             onClose={this.clearForm}
           />