+++ test/bigtest/interactors/block-modal.js
@@ -2,6 +2,7 @@ import {
   interactor,
   isPresent,
   collection,
+  scoped,
 } from '@bigtest/interactor';
 
 @interactor class BlockModalInteractor {
@@ -9,6 +10,7 @@ import {
 
   modalPresent = isPresent('[data-test-block-modal]');
   modalMessage = collection('[data-test-block-message]');
+  closeButton = scoped('[data-test-close-patron-block-modal]');
 
   whenBlockModalLoaded() {
     return this.when(() => this.modalPresent);