+++ test/bigtest/interactors/error-modal.js
@@ -1,4 +1,5 @@
 import {
+  collection,
   interactor,
   scoped,
   text,
@@ -7,9 +8,11 @@ import {
 @interactor class ErrorModal {
   static defaultScope = '[data-test-error-modal]';
 
-  content = text('[class^="modalContent---"] p');
+  content = text('[class^="modalContent---"] p:first-child');
   overrideButton = scoped('[data-test-override-button]');
   closeButton = scoped('[data-test-close-button]');
+  errorsList = scoped('[data-test-error-item]');
+  errorsListCollection = collection('[data-test-error-item]');
 }
 
 export default ErrorModal;