+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Proxy not allowed to checkout items for sponsor if proxy has patron block. Refs UICHKOUT-673.
 * Show confirmation modal when scanning an item with one of the new statuses (Long missing, In process (non-requestable), Restricted, Unavailable, Unknown). Refs UICHKOUT-671.
 * Increment `@folio/stripes-cli` to `v2`. Refs UICHKOUT-692.
+* Handle a full list of errors when check out fails. Refs UICHKOUT-679.
 
 ## [5.0.0](https://github.com/folio-org/ui-checkout/tree/v5.0.0) (2020-10-12)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v4.0.1...v5.0.0)